+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: allow only one redirect for a site and outbound path combination
 
 1.0.0 (2022-09-08)
 ==================